refactor(provider-node): reuse OnlineClients
