+++ provider-node/src/command.rs
@@ -13,8 +13,6 @@ use crate::{
 use clap::Parser;
 use std::sync::Arc;
 use std::time::Duration;
-use subxt::dynamic::At;
-use subxt::{dynamic::Value, OnlineClient, PolkadotConfig};
 use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt};
 
 /// Parse CLI arguments, initialize the node, and run the server.
@@ -125,15 +123,13 @@ pub async fn run() -> Result<(), Box<dyn std::error::Error>> {
     let _agreement_handle =
         start_agreement_coordinator(&cli, chain_client.as_ref(), state.clone()).await;
 
-    // Sync on-chain multiaddr with actual bind address (requires signing key)
-    if let Some(seed) = &seed {
-        sync_multiaddr_on_chain(
-            &cli.rpc.chain_rpc,
-            seed,
-            &state.provider_id,
-            &cli.rpc.bind_addr,
-        )
-        .await;
+    // Sync on-chain multiaddr with actual bind address. Reuses the chain
+    // client connected above, so this only runs when that connection succeeded
+    // (which also implies a signing key was provided).
+    if let Some(chain_client) = &chain_client {
+        chain_client
+            .sync_multiaddr(&state.provider_id, &cli.rpc.bind_addr)
+            .await;
     }
 
     tracing::info!("Starting storage provider node on {}", cli.rpc.bind_addr);
@@ -254,175 +250,6 @@ async fn start_agreement_coordinator(
         Err(e) => {
             tracing::error!("Failed to start agreement coordinator: {}", e);
             None
-        }
-    }
-}
-
-/// Convert a bind address (e.g. "0.0.0.0:3333") to a multiaddr string (e.g. "/ip4/127.0.0.1/tcp/3333").
-fn bind_addr_to_multiaddr(bind_addr: &str) -> String {
-    let parts: Vec<&str> = bind_addr.split(':').collect();
-    let (host, port) = if parts.len() == 2 {
-        (parts[0], parts[1])
-    } else {
-        ("127.0.0.1", "3333")
-    };
-    // 0.0.0.0 isn't useful as a client-facing address
-    let host = if host == "0.0.0.0" { "127.0.0.1" } else { host };
-    format!("/ip4/{host}/tcp/{port}")
-}
-
-/// Ensure the on-chain multiaddr matches the actual bind address.
-/// If the provider is registered and the multiaddr differs, submit an update transaction.
-async fn sync_multiaddr_on_chain(chain_rpc: &str, seed: &str, provider_id: &str, bind_addr: &str) {
-    let expected_multiaddr = bind_addr_to_multiaddr(bind_addr);
-
-    let api = match OnlineClient::<PolkadotConfig>::from_url(chain_rpc).await {
-        Ok(api) => api,
-        Err(e) => {
-            tracing::warn!("Could not connect to chain for multiaddr sync: {}", e);
-            return;
-        }
-    };
-
-    // Read current on-chain provider info
-    let our_account: sp_core::crypto::AccountId32 =
-        match sp_core::crypto::Ss58Codec::from_ss58check(provider_id) {
-            Ok(a) => a,
-            Err(_) => {
-                tracing::warn!("Invalid provider SS58 address, skipping multiaddr sync");
-                return;
-            }
-        };
-    let our_bytes: [u8; 32] = our_account.into();
-
-    let storage_query = subxt::dynamic::storage(
-        "StorageProvider",
-        "Providers",
-        vec![Value::from_bytes(our_bytes)],
-    );
-
-    let result = match api.storage().at_latest().await {
-        Ok(s) => s.fetch(&storage_query).await,
-        Err(e) => {
-            tracing::warn!("Failed to query storage for multiaddr sync: {}", e);
-            return;
-        }
-    };
-
-    let provider_value = match result {
-        Ok(Some(v)) => v,
-        Ok(None) => {
-            tracing::info!("Provider not registered on chain yet, skipping multiaddr sync");
-            return;
-        }
-        Err(e) => {
-            tracing::warn!("Failed to fetch provider info: {}", e);
-            return;
-        }
-    };
-
-    // Extract multiaddr from the encoded provider storage entry.
-    // Decode to scale_value and extract the "multiaddr" field bytes.
-    let current = {
-        let decoded = match provider_value.to_value() {
-            Ok(v) => v,
-            Err(e) => {
-                tracing::warn!("Could not decode provider value: {}, skipping sync", e);
-                return;
-            }
-        };
-
-        // Use At trait to navigate the decoded value
-        let multiaddr_val = match decoded.at("multiaddr") {
-            Some(v) => v,
-            None => {
-                tracing::warn!("No multiaddr field in provider info, skipping sync");
-                return;
-            }
-        };
-
-        // Extract bytes from the value — could be a sequence of u8 values
-        fn extract_bytes<T>(val: &subxt::ext::scale_value::Value<T>) -> Vec<u8> {
-            use subxt::ext::scale_value::{Composite, Primitive, ValueDef};
-            match &val.value {
-                // Sequence/composite of individual byte values
-                ValueDef::Composite(Composite::Unnamed(items)) => items
-                    .iter()
-                    .filter_map(|item| match &item.value {
-                        ValueDef::Primitive(Primitive::U128(n)) => Some(*n as u8),
-                        _ => None,
-                    })
-                    .collect(),
-                // Some subxt versions encode Vec<u8> as a Primitive::U256 or similar
-                ValueDef::Primitive(Primitive::U128(n)) => {
-                    // Single byte
-                    vec![*n as u8]
-                }
-                _ => vec![],
-            }
-        }
-
-        let bytes = extract_bytes(multiaddr_val);
-        if bytes.is_empty() {
-            // Debug: log the actual value structure to understand the encoding
-            tracing::warn!(
-                "multiaddr decoded as empty, value structure: {:?}",
-                multiaddr_val
-            );
-        }
-        String::from_utf8_lossy(&bytes).to_string()
-    };
-
-    if current == expected_multiaddr {
-        tracing::info!(
-            "On-chain multiaddr matches bind address: {}",
-            expected_multiaddr
-        );
-        return;
-    }
-
-    tracing::info!(
-        "On-chain multiaddr mismatch: chain=\"{}\" actual=\"{}\", updating...",
-        current,
-        expected_multiaddr
-    );
-
-    // Create signer from seed
-    let uri: subxt_signer::SecretUri = match seed.parse() {
-        Ok(u) => u,
-        Err(e) => {
-            tracing::error!("Failed to parse seed for multiaddr update: {:?}", e);
-            return;
-        }
-    };
-    let signer = match subxt_signer::sr25519::Keypair::from_uri(&uri) {
-        Ok(kp) => kp,
-        Err(e) => {
-            tracing::error!("Failed to create keypair for multiaddr update: {:?}", e);
-            return;
-        }
-    };
-
-    let multiaddr_bytes = expected_multiaddr.as_bytes().to_vec();
-    let tx = subxt::dynamic::tx(
-        "StorageProvider",
-        "update_provider_multiaddr",
-        vec![Value::from_bytes(multiaddr_bytes)],
-    );
-
-    match api
-        .tx()
-        .sign_and_submit_then_watch_default(&tx, &signer)
-        .await
-    {
-        Ok(progress) => match progress.wait_for_finalized_success().await {
-            Ok(_) => {
-                tracing::info!("Multiaddr updated on-chain to: {}", expected_multiaddr)
-            }
-            Err(e) => tracing::error!("Multiaddr update tx failed: {}", e),
-        },
-        Err(e) => {
-            tracing::error!("Failed to submit multiaddr update: {}", e);
         }
     }
 }
